Music engine work.
- Tons of work on the music system. Just need to finish mute action.
- Updated a few variables within the message system.
The music system is a bigger undertaking than I thought, but I thankfully planned it out well. I just need to add a temporary mute feature, and it is ready to deploy. We will have to figure out things like volume settings in options later, but it will be very easy; just replacing numbers with variables in the completed music handler script.