Minimap system implemented!!!
This was an awesome milestone for me. I did not want to go with your standard Metroidvania map system on a grid, so I went with something original (or at least original in the sense that I did not outright copy and paste anything). The maps will be hand drawn based on the walls in the actual rooms. Some adjustments definitely need to be made, but the minimap is very flexible, simplistic, and gets the job done in a gorgeous way: https://imgur.com/a/5okesmK. The best part: it worked on the very first try! I still have more plans, but this is fully functional for the time being.
- Added WIP interpolated sprite for the center of the minimap to show player location in the room. Might remove it, since it is pointless to have it actually follow the accurate player's location, or maybe I am too lazy to program it!